riscv: riscv-013: fix receive_addr calculation fault
When calculating the receive_addr, it substracts the loop temporary
variable i by 2, which is an unsigned integer (size_t) starts at 0.

As an unsigned integer should not be subtracted by a bigger number,
fix this formula by preventing this substraction.

This fixes wrong receive_addr when size_t is not of the same size
as riscv_addr_t.

Change-Id: I76ff8fc3f1215317eb49391e5b4923384d629dc4
Signed-off-by: Icenowy Zheng <icenowy@aosc.io>